# Conflicts:
#	labs/models.py
#	labs/urls.py
#	labs/views.py
#	templates/labs/wastes.html

+++ labs/views.py
@@ -1,10 +1,10 @@
-from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
+from django.shortcuts import render, redirect, get_object_or_404
 
 from registration.forms import UpdateMyUserForm
+from registration.models import MyUser
 from .forms import WasteForm
 from .models import Waste, BookmarkedWaste
-from registration.models import MyUser
 
 
 @login_required
@@ -40,7 +40,8 @@ def user_stats(request):
 
 @login_required
 def user_wastes(request):
-    data = {'my_wastes': Waste.objects.filter(generator=request.user)}
+    data = {'my_wastes': Waste.objects.filter(generator=request.user),
+            'my_bookmarked_wastes': Waste.objects.filter(generator=request.user)}
 
     return render(request, 'labs/wastes.html', data)
 
@@ -54,7 +55,7 @@ def user_wastes_create(request):
 
         # preenchimento do campo "gerador" baseado no usuário que está logado.
         waste.generator = request.user
-
+        waste.status_update()
         waste.save()
         return redirect('user_wastes')
 
@@ -63,32 +64,56 @@ def user_wastes_create(request):
 
 def user_wastes_update(request, waste_id):
     waste = get_object_or_404(Waste, pk=waste_id)
-    form = WasteForm(request.POST or None, instance=waste)
-
-    if form.is_valid():
-        waste.save()
+    if waste.status == 'user_inventory':
+        form = WasteForm(request.POST or None, instance=waste)
+        if form.is_valid():
+            waste.status_update()
+            waste.save()
+            return redirect('user_wastes')
+        return render(request, 'labs/waste_form.html', {'waste_form': form})
+    else:
         return redirect('user_wastes')
-
-    return render(request, 'labs/waste_form.html', {'waste_form': form})
-
 
 def user_wastes_delete(request, waste_id):
     waste = get_object_or_404(Waste, pk=waste_id)
+    if waste.status == 'user_inventory':
+        if request.method == 'POST':
+            waste.delete()
+            waste.status_update()
+            return redirect('user_wastes')
 
-    if request.method == 'POST':
-        waste.delete()
+        return render(request, 'labs/waste_delete.html', {'this_waste': waste})
+    else:
         return redirect('user_wastes')
 
-    return render(request, 'labs/waste_delete.html', {'this_waste': waste})
+def user_wastes_duplicate(request, waste_id):
+    new_waste = Waste.objects.get(pk=waste_id)
+    new_waste.pk = None
+    new_waste.status = 'user_inventory'
+    new_waste.status_update()
+    new_waste.save()
+
+    return redirect('user_wastes')
 
 
 def user_wastes_bookmark(request, waste_id):
     waste = get_object_or_404(Waste, pk=waste_id)
 
     if request.method == 'POST':
-        bookmarked_waste = BookmarkedWaste.objects.create(waste)
+        bookmarked_waste = BookmarkedWaste.objects.create(
+            bookmarked_waste=waste)
         bookmarked_waste.save()
 
         return redirect('user_wastes')
 
-    return render(request, 'labs/waste_bookmark.html')
+    return render(request, 'labs/waste_bookmark.html', {'this_waste': waste})
+
+def user_wastes_ask_removal(request, waste_id):
+    waste = get_object_or_404(Waste, pk=waste_id)
+    print('oi')
+    if request.method == 'POST':
+        waste.status = 'waiting_removal'
+        waste.save()
+        print('ola' + waste.status)
+        return redirect('user_wastes')
+    return render(request, 'labs/waste_ask_removal.html', {'this_waste': waste})
